implemented Registration and Delete account Features

+++ lib/data/provider/user_provider.dart
@@ -6,6 +6,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../core/utils/show_notifier_snack_bar.dart';
 import '../../core/utils/validators.dart';
+import '../../presentation/screens/auth/login_screen.dart';
 import '../models/user_model.dart';
 import '../repositories/auth_repo.dart';
 import '../repositories/user_repo.dart';
@@ -46,8 +47,8 @@ class UserController extends ChangeNotifier {
       _isLoading = true;
       notifyListeners();
 
-      final nameError = validateProfileName(name);
-      final emailError = validateProfileEmail(email);
+      final nameError = validateName(name);
+      final emailError = validateEmail(email);
 
       if (nameError != null || emailError != null) {
         ShowNotifierSnackBar.showSnackBar(
@@ -102,14 +103,34 @@ class UserController extends ChangeNotifier {
   Future<void> deleteUser(BuildContext context) async {
     try {
       final token = await AuthRepository.retrieveToken();
+
+      if (token == null) {
+        ShowNotifierSnackBar.showSnackBar(
+          context,
+          'Unauthenticated',
+          const Color(0xFF626FFF),
+          const Color(0xFFFFFFFF),
+        );
+        return;
+      }
+
       await UserRepository.deleteUser(token);
 
       _currentUser = null;
       final prefs = await SharedPreferences.getInstance();
       await prefs.remove('currentUser');
 
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Account deleted successful!')),
+      Navigator.of(context).push(
+        MaterialPageRoute(
+          builder: (context) => const LoginScreen(),
+        ),
+      );
+
+      ShowNotifierSnackBar.showSnackBar(
+        context,
+        'Account deleted!',
+        const Color(0xFFFFFFFF),
+        const Color(0xFF626FFF),
       );
     } catch (e) {
       print('Account deleted failed: ${e.toString()}');
